biowebRest/requirement.txt
aioredis==1.3.1
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.0
astroid==2.11.5
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==23.3.0
bleach==5.0.1
bootstrap4==0.1.0
celery==5.0.5
certifi==2022.6.15
cffi==1.15.1
channels==3.0.3
channels-redis==2.4.2
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
comm==0.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.3
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
entrypoints==0.4
executing==1.2.0
factory-boy==3.0.1
Faker==13.12.1
fastjsonschema==2.16.2
fqdn==1.5.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==6.0.0
incremental==21.3.0
inflection==0.5.1
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
iso-639==0.4.5
isoduration==20.11.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kombu==5.2.4
lazy-object-proxy==1.7.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
msgpack==0.6.2
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.4
packaging==23.1
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.30
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
pyee==8.2.2
Pygments==2.14.0
PyJWT==2.4.0
pylint==2.14.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.4
pytz==2022.1
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
redis==3.5.3
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
widgetsnbextension==4.0.5
wrapt==1.14.1
zipp==3.11.0
zope.interface==5.4.0
